+++ internal/provider/names.go
@@ -259,6 +259,7 @@ const (
 	keyObjectIDs                                    = "object_ids"
 	keyObjectType                                   = "object_type"
 	keyObjectTypes                                  = "object_types"
+	keyObjects                                      = "objects"
 	keyOperation                                    = "operation"
 	keyOperations                                   = "operations"
 	keyOptionalConfig                               = "optional_config"